Move image picker classes to own package
