+++ core/src/main/java/at/shockbytes/dante/core/image/picker/RxLegacyImagePicker.kt
@@ -1,4 +1,4 @@
-package at.shockbytes.dante.core.image
+package at.shockbytes.dante.core.image.picker
 
 import android.net.Uri
 import androidx.fragment.app.FragmentActivity
